# backend/tests/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "Body_upload_document_api_v1_matters__matter_id__documents_upload_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_document_api_v1_matters__matter_id__documents_upload_post",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MatterCreateIn": {
        "properties": {
          "jurisdiction": {
            "default": "casablanca",
            "maxLength": 100,
            "title": "Jurisdiction",
            "type": "string"
          },
          "language": {
            "default": "ar",
            "maxLength": 10,
            "title": "Language",
            "type": "string"
          },
          "matter_type": {
            "default": "general",
            "maxLength": 100,
            "title": "Matter Type",
            "type": "string"
          },
          "title": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "title"
        ],
        "title": "MatterCreateIn",
        "type": "object"
      },
      "MatterDocumentOut": {
        "description": "One row of GET /matters/{matter_id}/documents (frozen contract).",
        "properties": {
          "chunk_count": {
            "title": "Chunk Count",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "doc_type": {
            "title": "Doc Type",
            "type": "string"
          },
          "document_id": {
            "title": "Document Id",
            "type": "integer"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "needs_review": {
            "title": "Needs Review",
            "type": "boolean"
          },
          "original_name": {
            "title": "Original Name",
            "type": "string"
          },
          "page_count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page Count"
          },
          "section_count": {
            "title": "Section Count",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "document_id",
          "original_name",
          "filename",
          "doc_type",
          "status",
          "needs_review",
          "chunk_count",
          "section_count",
          "page_count",
          "created_at"
        ],
        "title": "MatterDocumentOut",
        "type": "object"
      },
      "MatterOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "jurisdiction": {
            "title": "Jurisdiction",
            "type": "string"
          },
          "language": {
            "title": "Language",
            "type": "string"
          },
          "matter_type": {
            "title": "Matter Type",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "matter_type",
          "jurisdiction",
          "language"
        ],
        "title": "MatterOut",
        "type": "object"
      },
      "MatterUpdateIn": {
        "description": "Partial update: every field optional; only supplied fields are applied.",
        "properties": {
          "jurisdiction": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jurisdiction"
          },
          "language": {
            "anyOf": [
              {
                "maxLength": 10,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "matter_type": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matter Type"
          },
          "title": {
            "anyOf": [
              {
                "maxLength": 255,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "title": "MatterUpdateIn",
        "type": "object"
      },
...
      "SectionOut": {
        "properties": {
          "faithful_text": {
            "title": "Faithful Text",
            "type": "string"
          },
          "needs_review": {
            "title": "Needs Review",
            "type": "boolean"
          },
          "normalized_text": {
            "title": "Normalized Text",
            "type": "string"
          },
          "ocr_confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ocr Confidence"
          },
          "page_end": {
            "title": "Page End",
            "type": "integer"
          },
          "page_start": {
            "title": "Page Start",
            "type": "integer"
          },
          "parent_section_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Section Id"
          },
          "section_id": {
            "title": "Section Id",
            "type": "string"
          },
          "span_end": {
            "title": "Span End",
            "type": "integer"
          },
          "span_start": {
            "title": "Span Start",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "section_id",
          "parent_section_id",
          "title",
          "page_start",
          "page_end",
          "span_start",
          "span_end",
          "faithful_text",
          "normalized_text",
          "ocr_confidence",
          "needs_review"
        ],
        "title": "SectionOut",
        "type": "object"
      },
...
      "UploadOut": {
        "properties": {
          "blob_ref": {
            "title": "Blob Ref",
            "type": "string"
          },
          "doc_type": {
            "title": "Doc Type",
            "type": "string"
          },
          "document_id": {
            "title": "Document Id",
            "type": "integer"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "filename": {
            "title": "Filename",
            "type": "string"
          },
          "indexed_count": {
            "title": "Indexed Count",
            "type": "integer"
          },
          "matter_id": {
            "title": "Matter Id",
            "type": "integer"
          },
          "needs_review": {
            "title": "Needs Review",
            "type": "boolean"
          },
          "sections": {
            "items": {
              "$ref": "#/components/schemas/SectionOut"
            },
            "title": "Sections",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "version_no": {
            "title": "Version No",
            "type": "integer"
          }
        },
        "required": [
          "matter_id",
          "document_id",
          "version_no",
          "filename",
          "doc_type",
          "status",
          "needs_review",
          "blob_ref",
          "indexed_count",
          "error",
          "sections"
        ],
        "title": "UploadOut",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Moroccan Legal RAG",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/matters": {
      "get": {
        "description": "List all matters (single-user local app: no pagination needed).",
        "operationId": "list_matters_api_v1_matters_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MatterOut"
                  },
                  "title": "Response List Matters Api V1 Matters Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Matters",
        "tags": [
          "matters"
        ]
      },
      "post": {
        "description": "Create a matter row; returns its id for uploads/analysis/chat.",
        "operationId": "create_matter_api_v1_matters_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatterCreateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatterOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Matter",
        "tags": [
          "matters"
        ]
      }
    },
    "/api/v1/matters/{matter_id}": {
      "delete": {
        "description": "Hard-delete a matter and everything scoped to it, purging vectors/blobs.",
        "operationId": "delete_matter_api_v1_matters__matter_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "matter_id",
            "required": true,
            "schema": {
              "title": "Matter Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Matter Api V1 Matters  Matter Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Matter",
        "tags": [
          "matters"
        ]
      },
      "get": {
        "description": "Fetch one matter by id; 404 when it does not exist.",
        "operationId": "get_matter_api_v1_matters__matter_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "matter_id",
            "required": true,
            "schema": {
              "title": "Matter Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatterOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Matter",
        "tags": [
          "matters"
        ]
      },
      "patch": {
        "description": "Apply only the fields present in the body; 404 when the matter is absent.",
        "operationId": "update_matter_api_v1_matters__matter_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "matter_id",
            "required": true,
            "schema": {
              "title": "Matter Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatterUpdateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatterOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Matter",
        "tags": [
          "matters"
        ]
      }
    },
...
    "/api/v1/matters/{matter_id}/documents": {
      "get": {
        "description": "List a matter's documents (id ascending); 200 [] when there are none.",
        "operationId": "list_documents_api_v1_matters__matter_id__documents_get",
        "parameters": [
          {
            "in": "path",
            "name": "matter_id",
            "required": true,
            "schema": {
              "title": "Matter Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MatterDocumentOut"
                  },
                  "title": "Response List Documents Api V1 Matters  Matter Id  Documents Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Documents",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/matters/{matter_id}/documents/upload": {
      "post": {
        "description": "Store the immutable original, run the pipeline, persist provenance.",
        "operationId": "upload_document_api_v1_matters__matter_id__documents_upload_post",
        "parameters": [
          {
            "in": "path",
            "name": "matter_id",
            "required": true,
            "schema": {
              "title": "Matter Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_document_api_v1_matters__matter_id__documents_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Document",
        "tags": [
          "documents"
        ]
      }
    },
    "/api/v1/matters/{matter_id}/documents/{document_id}": {
      "delete": {
        "description": "Hard-delete one document: rows, evidence points, and blob originals.",
        "operationId": "delete_document_api_v1_matters__matter_id__documents__document_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "matter_id",
            "required": true,
            "schema": {
              "title": "Matter Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "document_id",
            "required": true,
            "schema": {
              "title": "Document Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Delete Document Api V1 Matters  Matter Id  Documents  Document Id  Delete",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Document",
        "tags": [
          "documents"
        ]
      }
    },
...
  }
}